Adjustments to registries for thread safety.
Provide a thread_safe_registry_via_id.  This class should arguably go
somewhere else, like cetlib.  However, that would involve bringing TBB
down all the way to cetlib.  I'm not sure we want to do that just yet.

In principle, the thread_safe_registry_via_id can be replaced with a
database.  However, until we're willing to go that route, this will
have to do.

This commit does not provide unit tests for the thread-safe registry.