Add MF_PLUGIN_PATH to CheckClassVersion env
From 2.2.0, MessageFacility has global plugin factories in
MessageLogger.cc. Thus any client of MF needs the dynamic loader
path set, or the library will throw, irrespective of whether the
client uses MF or not. This affects the run of CheckClassVersion
due to it loading libraries, and consequently failing due to
missing library paths.

Add MF_PLUGIN_PATH to the environment under which checkClassVersion
runs. Set it to any existing value only, as it can be blank. This
is compatible with older versions. It will not interfere with
any prexisting configurations.